+++ RevitAddin.Tasks.Example/Revit/App.cs
@@ -1,9 +1,7 @@
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
-using RevitAddin.Tasks.Example.Services;
 using ricaun.Revit.UI;
-using System;
-using System.Threading.Tasks;
+using ricaun.Revit.UI.Tasks;
 
 namespace RevitAddin.Tasks.Example.Revit
 {
@@ -12,35 +10,23 @@ namespace RevitAddin.Tasks.Example.Revit
     {
         private static RibbonPanel ribbonPanel;
         private static RevitTaskService revitTaskService;
-        public static IRevitTask RevitTask { get; private set; }
+        public static IRevitTask RevitTask => revitTaskService;
         public Result OnStartup(UIControlledApplication application)
         {
-            ribbonPanel = application.CreatePanel("RevitAddin.Tasks.Example");
-            ribbonPanel.CreatePushButton<Commands.Command>()
-                .SetLargeImage("Resources/Revit.ico");
-
             revitTaskService = new RevitTaskService(application);
             revitTaskService.Initialize();
 
-            RevitTask = new RevitTask(revitTaskService);
-
-            //var task = Task.Run(async () =>
-            //{
-            //    await Task.Delay(1000);
-            //    await RevitTask.Run(() => { Console.WriteLine("Like this video!"); });
-            //    await Task.Delay(1000);
-            //    await RevitTask.Run(() => { Console.WriteLine("Subscribe to my channel!"); });
-            //});
-
+            ribbonPanel = application.CreatePanel("RevitAddin.Tasks.Example");
+            ribbonPanel.CreatePushButton<Commands.Command>()
+                .SetLargeImage("Resources/Revit.ico");
             return Result.Succeeded;
         }
 
         public Result OnShutdown(UIControlledApplication application)
         {
+            revitTaskService?.Dispose();
+
             ribbonPanel?.Remove();
-
-            revitTaskService.Dispose();
-
             return Result.Succeeded;
         }
     }