+++ RevitAddin.Tasks.Example/Revit/Commands/Command.cs
@@ -1,9 +1,7 @@
 using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
-using RevitAddin.Tasks.Example.Services;
-using RevitAddin.Tasks.Example.Views;
-using System;
+using ricaun.Revit.UI.Tasks;
 using System.Threading.Tasks;
 
 namespace RevitAddin.Tasks.Example.Revit.Commands
@@ -15,7 +13,7 @@ namespace RevitAddin.Tasks.Example.Revit.Commands
         {
             UIApplication uiapp = commandData.Application;
 
-            new MainView(App.RevitTask).Show();
+            new Views.MainView(App.RevitTask).Show();
 
             return Result.Succeeded;
         }