+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] / 2024-03-12
+- Add `ricaun.Revit.UI.Tasks` package.
+- Update `MainView` to select element image.
+- ![Tasks and WPF](https://github.com/ricaun-io/RevitAddin.Tasks.Example/assets/12437519/7be8ed93-f5c5-41c0-a75e-8c514f5be10a)
+
 ## [1.0.0] / 2024-01-16
 - First Release
 - `IRevitTask`, `RevitTask` and `RevitTaskService`
 
 [vNext]: ../../compare/1.0.0...HEAD
+[2.0.0]: ../../compare/1.0.0...2.0.0
 [1.0.0]: ../../compare/1.0.0